GridApplicationDelegate: make keyboard-accessible
The home page in Discover shows Cards for a series of categories,
including Most Popular, Editor's Choice, etc. These can only interacted
with using the mouse; there's no way to access them using the keyboard,
and users relying on screen readers for discovery will not be aware that
they exist at all.

This change makes the Cards part of the tab chain, and sets the a11y
attached properties so that the application name and a suitable type
(Link, as activating the cards opens the corresponding application page
in Discover) are read out.