BrowsingPage: center view on card/button if outside
The Home page of Discover shows a series of categories containing links
to applications, and for some of them a button that will open the
category browser. If tab navigation is used to reach these from the
keyboard and the window is small enough that scrolling is necessary to
display all of them, the CardsLayout will not automatically position
the Flickable to keep the newly focused item visible. This makes
keyboard navigation inconvenient (at least without relying on a screen
reader). CardsLayout is not a view, and does not have convenience
functions for this purpose.

This change updates the Flickable of the ScrollingPage, such that if the
card or button would not be completely visible, the scrolling position
is set to center the newly focused item as well as possible given the
bounds of the Flickable.