ApplicationsListPage: fix unnecessary horizontal scrolling
The content comes from a Kirigami.CardsListView, which brings its own
horizontal margins that interfere with default paddings from
DiscoverPage/Kirigami.ScrollablePage and cause the page to be slightly
horizontally scrollable.

We need to make sure that there are no overlapping margins, which can be
done by setting the CardsListView to the page's flickable property and
nulling out the page's top and bottom padding values. UX remains the same
other than the bug being fixed.

BUG: 477493
FIXED-IN: 6.3.0