ApplicationPage: Fix excessive horizontal flicking
Explicitly set the preferred layout width for the selectable labels
to the actual width after wrapping (`contentWidth`), instead of the
default `implicitWidth`.

For selectable labels, the `implicitWidth` is the full width of the
text before wrapping. When his value is big enough, it bubbles up
the layout chain and ends up allowing to horizontally scroll the
view to the right.

There is no binding loop, since the actual width calculation already
uses `fillWidth`.