Remove AbstractResourcesBackend::isFetching
The concept of fetching doesn't translate very well to our abstraction
and nowadays is better served by streams.
Streams are supposed to either fail or delay when fetching, we remove a
global state of fetching-ness in the system where the app is not useful.

It makes for a more fluid UX.