Fix nonfree app license name colors
We use Kirigami.UrlButton here and set it to disabled when there's no
URL for the license. However in this case Kirigami.UrlButton always used
the disabled text color.

This isn't a wrong behavior for the library component to have, but it
doesn't suit our needs here. We want to use the appropriate color even
for disabled links, and we want to do it without simply overriding the
base color.

This commit fixes the issue by undoing some behaviors of the library
component in places where they're not needed.