flatpak: do not mark eol runtimes for uninstall
this breaks transactions with EOL runtimes that are still used.
specifically we'll add an uninstall operation to the transaction but
that operation will fail when the runtime is still used by apps (which
too may be EOL, mind).

instead we just discard the EOL callback and keep the runtime installed.
the runtime will eventually get removed when the last user gets removed
by the user.