KNSBackend: Fix Updates getting stuck for wrong KNS entries
When searching for updates on a KNS entry that points to a wrong or
non-existent category, the provider indicates the error by emitting a
`signalErrorCode` signal.

In this case, let's finish the search job (`ResultStream`), or it will
be kept hanging forever.

This fixes discover Updates page being stuck at the "fetching" state,
as the `StandardBackendUpdater` that launches the search jobs need it
to either succeed (`resourcesFound`) or get `destroyed` to change the
progress state.

BUG: 500513
FIXED-IN: 6.3.3