Show transaction progress at point of use
Right now when you install an app from the application page, the
progress indicator appears in the bottom-left corner of the window,
which is the opposite corner from where the user's pointer is located
and where their eyeballs are looking. It's easy to miss.

This commit changes that by moving the progress information into the
location where the user just clicked and is still looking. The old
transaction progress overview is hidden while there's only one
in-progress transaction and it's already visible on the page.

FEATURE: 475845
FIXED-IN: 6.4.0