discover/qml: improve UpdatesPage accessibility
Discover's Updates page is currently rather awkward to use with the
keyboard: the only thing that gets focus are the checkmarks, but they do
not have any focus indicator when they do not have their own label, so
it is not clear where focus is. For screen reader users it is a bit more
obvious that a checkbox is focused, but no text is announced so it's
unclear what the checkbox is for.

This change makes the list work similar to other lists: the whole
delegate indicates focus, and the user can press space directly on the
delegate to toggle the checkbox state (as in e.g. kdeconnect's plugins
list). Screen readers announce the text and subtitle, as well as the
number of list entries when entering the list. The checkboxes no longer
individually take focus to simplify interaction. Return now extends the
delegate to show the extra information (as an alternative to holding
Alt, which still works).


(cherry picked from commit 7da50005ae616881b1042a9cf1c95520907e7c1e)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>