rpm-ostree: Fix rpm-ostree driver registration
Properly initilize the boolean responsible for trigerring the logic to
register Discover as an update driver to the rpm-ostree daemon.

This will keep the rpm-ostree daemon running (but mostly idle) as long
as the Discover main window is open, and will avoid re-loading the list
of deployments and update every ~60 seconds, when rpm-ostree would have
exited due to being idle, releasing the name on the bus.

Fixes: https://bugs.kde.org/show_bug.cgi?id=480947

BUG: 480947