FlatpakTransactionThread: Copy job to new ref after rebase
When rebasing, the transaction still has the reference of the old item.
When rebasing is done and we know the new ref, the reference we ask for
during the new_operation_cb sets the new reference as current. This
causes an assert to fail the transaction, since the transaction
has the reference for the old item, even the job is still the same job.

Possible fix for this is to just copy the job and insert new ref as key
to the m_jobTransactionsByRef map, with the copied job: The job is
the same job for both the old and new reference.

BUG:505245


(cherry picked from commit 1c6ff6894e551a53c387d9f9a9f0429f63442e50)

Co-authored-by: Akseli Lahtinen <akselmo@akselmo.dev>