ApplicationsListPage: give listview default focus
The ListView on ApplicationsListPage does not have default focus, so if
something like running a search places focus on the results page, focus
gets stuck on the page itself rather than being passed down, and arrow
keys scroll the list, rather than select items.

Instead, mark the ListView as focus: true, so it'll have default focus
in the page's focusscope.

BUG: 505551
FIXED-IN: 6.4.1