discover: include Remove button in tab chain
The Remove button was excluded from the tab chain many years ago in
07aaf62e6ac932b4fed757208db36bd882c45e4e. This makes it impossible to
get to the button using tab navigation, which wasn't a problem when the
button was used as an extra on ListView cards that users could access
through the application's full page. But now that we use it also on that
page, we need to have it included or the button isn't accessible either.

Instead, we include the button in the tab chain when it's used on the
app's page. We still leave it out for ListViews, as having potentially
hundreds of delegates included is very cumbersome (and doesn't even work
well), though we may want to adjust this in the future to include the
button if and only if it is on the currently selected delegate.

BUG: 505544
FIXED-IN: 6.4.1


(cherry picked from commit 903b3745017cf3a8bc6674702117770f1f58f072)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>