sourcespage: fix delegate referencing the wrong `name`
in the switch to required properties supposedly the reference to `name`
got missed. unfortunately there is another `name` defined on the page
itself, so we ended up using the page name instead of the model name.

require that the model supplies us with a name role and use that.

BUG: 505700


(cherry picked from commit cf3e2bb650cba2ded4ecef3615c26e1dd6913e11)

Co-authored-by: Harald Sitter <sitter@kde.org>