sysupdate: Fix ending of updates
We'd update the version to "" (as indicated in the comment there). When
successful, set the new installed version of the resource to the one
that was available prior to the update.

Otherwise it would end up in an odd state where it seems it hasn't
updated but it actually has.

BUG: 502165