kcm: Restore runtime check for OSTree backend
It's possible to build discover with the OSTree backend while running on
a system that doesn't actually use OSTree. This would mean there's no
option for selecting how system updates work even though the system
supports choosing that, as the current logic is purely tied to a compile
time check. This restores the runtime check if the OSTree backend was
built. The systemd case is still based on a compile-time check as I
don't know how to check that at runtime.

BUG: 510072