Use safer org.kde.Shutdown to shut down and restart
Currently we use org.freedesktop.login1, which will not let apps prompt
the user to save changes. This can lead to data loss.

Ideally org.freedesktop.login1 would be plumbed through our session
stuff so apps can prompt, but that's quite a ways away. In the meantime,
use org.kde.Shutdown to resolve the issue.

BUG: 499260
FIXED-IN: 6.4.6