rpm-ostree: Do not fetch tags with skopeo
Tell skopeo to not fetch the list of container tags available in the
registry. We don't need it as we only look at the version of the tag
that we are requesting.

This significantly speeds up updates checks for the Bootable Container
path.

```
$ time skopeo inspect docker://quay.io/travier/fedora-kinoite:testing > /dev/null
0,05s user 0,02s system 2% cpu 2,749 total

$ time skopeo inspect --no-tags docker://quay.io/travier/fedora-kinoite:testing > /dev/null
0,04s user 0,02s system 5% cpu 1,214 total

$ time skopeo inspect docker://quay.io/fedora-ostree-desktops/kinoite:42 > /dev/null
0,06s user 0,03s system 1% cpu 5,810 total

$ time skopeo inspect --no-tags docker://quay.io/fedora-ostree-desktops/kinoite:42 > /dev/null
0,05s user 0,02s system 4% cpu 1,598 total
```

(cherry picked from commit a0ec6f487423c60122ac959ed34fd009a8f9b494)