Add sorting to the progress view
Makes sure that the transaction that is active at any given time is
listed on top and easily reachable to the user.
This is especially important since we removed the global entry on top of
the whole update. Flatpak will take all of the things that need updating
and update them at its own pace, so we better make it easy to the user
when there's a bunch.

Also adds some code to test in the dummy backend.